Refactored Tables to seperate components
